Moves scss from pages to styles/pages
