+++ app/styles/pages/error.scss
@@ -1,4 +1,4 @@
-@use '../styles/breakpoints.scss' as *;
+@use '../breakpoints.scss' as *;
 
 .error-page-container {
   display: flex;
