+++ app/styles/pages/home.scss
@@ -1,4 +1,4 @@
-@use '../styles/pxToRem.scss' as *;
+@use '../pxToRem.scss' as *;
 
 .home-container {
   display: flex;
