+++ app/styles/pages/monitor.scss
@@ -1,4 +1,4 @@
-@use '../styles/pxToRem.scss' as *;
+@use '../pxToRem.scss' as *;
 
 .monitor-container {
   display: flex;
