+++ app/styles/pages/notifications.scss
