+++ app/styles/pages/register.scss
@@ -1,4 +1,4 @@
-@use '../styles/pxToRem.scss' as *;
+@use '../pxToRem.scss' as *;
 
 .auth-form-container {
   width: 100vw;
